Refactor files locations
