+++ kudos/Presentation/Carousel/AccomplishmentDetailView.swift
@@ -1,4 +1,3 @@
-import SwiftData
 import SwiftUI
 
 struct AccomplishmentDetailView: View {
