+++ kudos/Presentation/Utilities/Confetti/Shapes/RoundedCross.swift
