+++ kudos/Presentation/Utilities/Confetti/Shapes/SlimRectangle.swift
