--- kudos/Presentation/Utilities/Constants/Visual.swift
@@ -1,8 +1,0 @@
-import SwiftUI
-
-enum Visual {
-    // Minimum scale factors
-    static let minimumScaleFactorSmall: CGFloat = 0.7
-    static let minimumScaleFactorMedium: CGFloat = 0.8
-}
-
