+++ kudos/Presentation/Utilities/Localization/Copies.swift
@@ -274,9 +274,6 @@ enum Copies {
         static func introTitle(year: Int) -> String {
             "wrapped_intro_title".localized.replacingOccurrences(of: "{year}", with: "\(year)")
         }
-        static func introSubtitle(count: Int) -> String {
-            "wrapped_intro_subtitle".localized.replacingOccurrences(of: "{count}", with: "\(count)")
-        }
         static var outroTitle: String { "wrapped_outro_title".localized }
         static var outroSubtitle: String { "wrapped_outro_subtitle".localized }
         static var close: String { "wrapped_close".localized }
