+++ kudos/Services/TipJarStore.swift
